implementacoes finais e acabou!!!!

+++ app/Controllers/CadastroAlunoController.php
@@ -6,18 +6,27 @@ use App\Models\AlunoModel;
 use App\Controllers\LoginController;
 use App\Controllers\BaseController;
 use CodeIgniter\HTTP\ResponseInterface;
+use App\Models\HistoricoModel;
 
 class CadastroAlunoController extends BaseController
 {
     private $AlunoModel;
     private $LoginController;
+    private $HistoricoModel;
+
     public function __construct(){
         $this->AlunoModel = new AlunoModel();
         $this->LoginController = new LoginController();
+        $this->HistoricoModel = new HistoricoModel();
     }
     public function index()
     {
-        return view('cadastro_aluno/cadastro-aluno');
+        $this->verificarBibliotecario();
+
+        $Aluno = $this->listar();
+        echo view('cadastro_aluno/cadastro-aluno', [
+            'Aluno' => $Aluno
+        ]);
     }
 
     public function listar(){
@@ -25,9 +34,24 @@ class CadastroAlunoController extends BaseController
     }
 
     public function salvar(){
+        if ($redirect = $this->verificarBibliotecario()) {
+            return $redirect;
+        }
+
         $Aluno = $this->request->getPost();
         $this->LoginController->salvar($Aluno);
         $this->AlunoModel->save($Aluno);
+
+        // Registrar no histórico
+        $this->HistoricoModel->registrarAcao(
+            'CADASTRO_ALUNO',
+            $Aluno['cpf'] ?? null,
+            $Aluno['nome'] ?? 'Desconhecido',
+            null,
+            null,
+            'Aluno cadastrado no sistema'
+        );
+
         return redirect()->to('/bibliotecario/cadastro_aluno');
     }
 
@@ -37,11 +61,19 @@ class CadastroAlunoController extends BaseController
     }
 
     public function editar($cpf){
+        if ($redirect = $this->verificarBibliotecario()) {
+            return $redirect;
+        }
+
         $Aluno = $this->procurar($cpf);
         echo view('cadastro_aluno/edit', ['Aluno' => $Aluno]);
     }
 
     public function excluir($cpf){
+        if ($redirect = $this->verificarBibliotecario()) {
+            return $redirect;
+        }
+
         $this->AlunoModel->delete($cpf);
         return redirect()->to('/bibliotecario/cadastro_aluno');
     }
