+++ app/Controllers/LivroController.php
@@ -3,17 +3,22 @@
 namespace App\Controllers;
 
 use App\Models\LivroModel;
+use App\Controllers\TomboController;
 use App\Controllers\BaseController;
 use CodeIgniter\HTTP\ResponseInterface;
 
 class LivroController extends BaseController
 {
     private $LivroModel;
+    private $TomboController;
     public function __construct(){
         $this->LivroModel = new LivroModel();
+        $this->TomboController =  new TomboController();
     }
     public function index()
     {
+        $this->verificarBibliotecario();
+
         return view('livro/cadastro-livro');
     }
 
@@ -22,7 +27,14 @@ class LivroController extends BaseController
     }
 
     public function salvar(){
+        if ($redirect = $this->verificarBibliotecario()) {
+            return $redirect;
+        }
+
         $Livro = $this->request->getPost();
+        $Dados = $this->TomboController->procurar($Livro['registro']);
+        $Livro['titulo'] = $Dados['titulo'];
+        $Livro['autor'] = $Dados['autor'];
         $this->LivroModel->save($Livro);
         return redirect()->to('/bibliotecario/livros');
     }
@@ -33,11 +45,19 @@ class LivroController extends BaseController
     }
 
     public function editar($id){
+        if ($redirect = $this->verificarBibliotecario()) {
+            return $redirect;
+        }
+
         $Livro = $this->procurar($id);
         echo view('livro/edit', ['Livro' => $Livro]);
     }
 
     public function excluir($id){
+        if ($redirect = $this->verificarBibliotecario()) {
+            return $redirect;
+        }
+
         $this->LivroModel->delete($id);
         return redirect()->to('/bibliotecario/livros');
     }
